Refactor: separate filters and helpers into discrete files/functions.
