+++ src/filters.ts
@@ -1,35 +1,42 @@
-import { FaExternalLinkAlt, FaSteamSymbol } from "react-icons/fa";
 import { Game } from "./entities/Game";
-import {
-  DisplayTabs,
-  FriendsOfFenAster,
-  RecsToType,
-  SortByType,
-} from "./store/gameListSlice";
+import { getPriceFloat, recByFriend } from "./helpers";
+import { DisplayTabs, RecsToType, SortByType } from "./store/gameListSlice";
 import { store } from "./store/store";
-
-export const enum Storefronts {
-  Steam,
-  Epic,
-  Other,
-}
-
-export const recByFriend = (game: Game): boolean =>
-  FriendsOfFenAster.includes(game.recBy);
 
 export const getFilteredGames = (games: Game[]): Game[] => {
   const {
-    showCompleted,
-    showRecsTo,
     showRecsBy,
     showOnlyFriends,
     searchTerm,
-    sortBy,
     sortResultsDescending,
     currentDisplayTab,
   } = store.getState().gameList;
 
   let filteredGames = [...games];
+  filteredGames = filterByCompleted(filteredGames);
+  filteredGames = sortGamesBy(filteredGames);
+  filteredGames = filterGamesByDisplayTab(currentDisplayTab, filteredGames);
+
+  if (showRecsBy.length > 0)
+    filteredGames = filteredGames.filter((game) =>
+      showRecsBy.includes(game.recBy)
+    );
+
+  if (showOnlyFriends)
+    filteredGames = filteredGames.filter((game) => recByFriend(game));
+
+  if (searchTerm.length > 1)
+    filteredGames = filteredGames.filter((game) =>
+      game.title.toLowerCase().includes(searchTerm.toLowerCase())
+    );
+
+  if (!sortResultsDescending) filteredGames = filteredGames.reverse();
+
+  return filteredGames;
+};
+
+const filterByCompleted = (filteredGames: Game[]): Game[] => {
+  const { showCompleted, showRecsTo } = store.getState().gameList;
 
   if (!showCompleted)
     filteredGames = filteredGames.filter((game) => !game.wasCompleted);
@@ -51,18 +58,11 @@ export const getFilteredGames = (games: Game[]): Game[] => {
       break;
   }
 
-  if (showRecsBy.length > 0)
-    filteredGames = filteredGames.filter((game) =>
-      showRecsBy.includes(game.recBy)
-    );
+  return filteredGames;
+};
 
-  if (showOnlyFriends)
-    filteredGames = filteredGames.filter((game) => recByFriend(game));
-
-  if (searchTerm.length > 1)
-    filteredGames = filteredGames.filter((game) =>
-      game.title.toLowerCase().includes(searchTerm.toLowerCase())
-    );
+const sortGamesBy = (filteredGames: Game[]): Game[] => {
+  const { sortBy } = store.getState().gameList;
 
   switch (sortBy) {
     case SortByType.title:
@@ -88,10 +88,6 @@ export const getFilteredGames = (games: Game[]): Game[] => {
       );
       break;
   }
-
-  filteredGames = filterGamesByDisplayTab(currentDisplayTab, filteredGames);
-
-  if (!sortResultsDescending) filteredGames = filteredGames.reverse();
 
   return filteredGames;
 };
@@ -124,51 +120,4 @@ const filterGamesByDisplayTab = (
   }
 
   return games;
-};
-
-export const getPriceFloat = (game: Game): number => {
-  switch (game.msrp.toLowerCase()) {
-    case "free":
-      return 0.0;
-    case "varies":
-      return 0.1;
-    case "n/a":
-      return 0.2;
-    default:
-      return parseFloat(game.msrp.substring(1, game.msrp.length));
-  }
-};
-
-export const detectStorefront = (game: Game) => {
-  if (game.storeLink?.includes("steampowered")) return Storefronts.Steam;
-  return Storefronts.Other;
-};
-
-export const getLinkIcon = (game: Game) =>
-  detectStorefront(game) === Storefronts.Steam ? (
-    <FaSteamSymbol />
-  ) : (
-    <FaExternalLinkAlt />
-  );
-
-export const extractSteamId = (game: Game): string | null => {
-  if (detectStorefront(game) !== Storefronts.Steam) return null;
-  if (!game.storeLink) return null;
-  const linkArr = game.storeLink.split("/app/");
-  if (linkArr[1] === undefined) return null;
-  const steamIdArr = linkArr[1].split("/");
-  return steamIdArr[0];
-};
-
-export const getGameImageUrl = (game: Game): string | null => {
-  const steamId = extractSteamId(game);
-  if (detectStorefront(game) === Storefronts.Steam)
-    return `https://cdn.cloudflare.steamstatic.com/steam/apps/${steamId}/header.jpg`;
-  return null;
-};
-
-export const calculateDiscount = (game: Game) => {
-  return (getPriceFloat(game) * (1 - game.discounts.discountPercent)).toFixed(
-    2
-  );
 };
